+++ tokogae-data/tokogae-data-event/tokogae-data-event-service/src/main/java/com/tokogae/data/event/service/impl/FoodItemServiceImpl.java
@@ -26,7 +26,7 @@ public class FoodItemServiceImpl extends FoodItemServiceBaseImpl {
 
 	public FoodItem addFoodItem(
 			long subjectId, long occurDay, int occurDaySegment, long occurTime,
-			String name, String locationOfOrigin, String brand, int quantity,
+			String name, String locationOfOrigin, String brand, double quantity,
 			String quantityUnit)
 		throws PortalException {
 
@@ -41,7 +41,7 @@ public class FoodItemServiceImpl extends FoodItemServiceBaseImpl {
 
 	public FoodItem updateFoodItem(
 			long foodItemId, long occurDay, int occurDaySegment, long occurTime,
-			String name, String locationOfOrigin, String brand, int quantity,
+			String name, String locationOfOrigin, String brand, double quantity,
 			String quantityUnit)
 		throws PortalException {
